new shooter map to 13

+++ src/main/java/frc/robot/commands/managers/CalibrationControl.java
@@ -57,7 +57,7 @@ public class CalibrationControl {
         });
     }
 
-    public Command incrementDistance(int increment) {
+    public Command incrementDistance(double increment) {
         return Commands.runOnce(() -> {
             distance += increment; 
             logDistance();
